Added new class called AccountType to be selected when an user is adding
a new movement. The method to register a new movement is almost ready.
Finally, I implement a new binary search tree in Account class, to
manage the movement in every account type and the tree in CashManager is
there to manage the movements to show them to the user in main page