The labels in main page were initialized, the table view in main page
too. Now when the user choose a type of account the combo box in
movement page to select an account display onlye the account that
corresponds with that type account selected. I am working to initialize
the rest of the tables view. Finally, the method to add a new movement
was modified, I add new validations.